pyproject.toml
[project]
name = "crewai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.102.0",
    "crewai-tools>=0.36.0",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
]
